Fix Windows breakage caused by #307
This is a followup to #307. I broke Windows with my change there, because single quotes can't be used in the command prompt as on Unix systems.

This is not ideal because `classpath` could be interpreted by the shell. Better would be to not run the command with `shell=True` at all and instead provide a list to `Popen` to let the subprocess module handle escaping of paths and such.